chore: css details
